Fix `BayesianEstimator` tests
