+++ tests/estimator/bayesian.rs
@@ -3,8 +3,8 @@ mod tests {
     use approx::*;
     use causal_hub_next::{
         data::CategoricalData,
-        distribution::{Distribution, BayesianEstimator},
-        estimator::Estimator,
+        distribution::Distribution,
+        estimator::{Estimator, BE},
     };
     use ndarray::prelude::*;
 
@@ -25,7 +25,7 @@ mod tests {
         ];
         let data = CategoricalData::new(&variables, values);
 
-        let estimator = BayesianEstimator::new(&data, 1.0);
+        let estimator = BE::new(&data, 1.0);
 
         // P(A)
         let distribution = estimator.fit(0, &[]);
@@ -36,11 +36,11 @@ mod tests {
             .values()
             .all(|x| x.iter().eq(["no", "yes"])));
 
-        assert_eq!(
+        assert_relative_eq!(
             distribution.parameters(),
             &array![
                 // A: no, yes
-                [0.5, 0.5]
+                [0.5714285714285714, 0.42857142857142855]
             ]
         );
 
@@ -48,7 +48,7 @@ mod tests {
         assert_eq!(distribution.sample_size(), Some(5));
         assert_relative_eq!(
             distribution.sample_log_likelihood().unwrap(),
-            -3.4657359027997265
+            -3.3734430845806758
         );
 
         assert_eq!(
@@ -59,7 +59,7 @@ mod tests {
                 "| ---- | ---- |\n",
                 "| no   | yes  |\n",
                 "| ---- | ---- |\n",
-                "| 0.50 | 0.50 |\n",
+                "| 0.57 | 0.43 |\n",
                 "---------------\n",
             )
         );
@@ -73,14 +73,14 @@ mod tests {
             .values()
             .all(|x| x.iter().eq(["no", "yes"])));
 
-        assert_eq!(
+        assert_relative_eq!(
             distribution.parameters(),
             &array![
                 // A: no, yes
                 [0.6666666666666666, 0.3333333333333333], // B: no, C: no
                 [0.6666666666666666, 0.3333333333333333], // B: no, C: yes
                 [0.3333333333333333, 0.6666666666666666], // B: yes, C: no
-                [0.4, 0.6]  // B: yes, C: yes
+                [0.5, 0.5]                                // B: yes, C: yes
             ]
         );
 
@@ -88,7 +88,7 @@ mod tests {
         assert_eq!(distribution.sample_size(), Some(5));
         assert_relative_eq!(
             distribution.sample_log_likelihood().unwrap(),
-            -1.6094379124341003
+            -2.602689685444384
         );
 
         assert_eq!(
@@ -102,7 +102,7 @@ mod tests {
                 "| no   | no   | 0.67 | 0.33 |\n",
                 "| no   | yes  | 0.67 | 0.33 |\n",
                 "| yes  | no   | 0.33 | 0.67 |\n",
-                "| yes  | yes  | 0.40 | 0.60 |\n",
+                "| yes  | yes  | 0.50 | 0.50 |\n",
                 "-----------------------------\n",
             )
         );
@@ -126,32 +126,9 @@ mod tests {
         ];
         let data = CategoricalData::new(&variables, values);
 
-        let estimator = BayesianEstimator::new(&data, 1.0);
+        let estimator = BE::new(&data, 1.0);
 
         // P(A | A, C)
         let _distribution = estimator.fit(0, &[0, 2]);
-    }
-
-    #[test]
-    #[should_panic(expected = "Marginal counts must be non-zero.")]
-    fn test_non_zero_counts() {
-        let variables = vec![
-            ("A", vec!["no", "yes"]),
-            ("B", vec!["no", "yes"]),
-            ("C", vec!["no", "yes"]),
-        ];
-        let values = array![
-            // A, B, C
-            [0, 0, 0],
-            [0, 0, 1],
-            [0, 1, 1],
-            [1, 1, 1]
-        ];
-        let data = CategoricalData::new(&variables, values);
-
-        let estimator = BayesianEstimator::new(&data, 1.0);
-
-        // P(A | B, C)
-        let _distribution = estimator.fit(0, &[1, 2]);
     }
 }
