+++ tests/estimator/mod.rs
@@ -1,1 +1,2 @@
+mod bayesian;
 mod maximum_likelihood;
